+++ core/scripts/installDependencies.sh
@@ -1,60 +1,34 @@
-#!/bin/bash
-set -x
+#! /bin/bash
+#set -x
 
-# Variables
-PROGRESS_FILE=/tmp/jeedom/Abeille/dependency_abeille_in_progress
+###
+### MAIN
+###
 
-# Functions
-
-function arret
-{
-  echo
-  echo "Avancement: 99% ---------------------------------------------------------------------------------------------------> ;-) "
-  echo
-
-  echo "Fin installation des dépendances"
-
-  echo
-  echo "Avancement: 100% ---------------------------------------------------------------------------------------------------> FIN"
-  echo
-
-  echo 100 > ${PROGRESS_FILE}
-  sleep 3
-  #suppression du decompte d'installation
-  rm ${PROGRESS_FILE}
-
-}
-
-function arretSiErreur
-{
-  echo
-  echo "***************"
-  echo $1
-  echo "***************"
-  echo
-  arret
-  exit 1
-}
-
-# MAIN
-echo "Début d'installation des dépendances"
-
-### INIT
+PROGRESS_FILE=/tmp/jeedom/Abeille/dependencies_progress
 if [ ! -z $1 ]; then
-	PROGRESS_FILE=$1
+    PROGRESS_FILE=$1
 fi
+
 touch ${PROGRESS_FILE}
-
 echo 0 > ${PROGRESS_FILE}
-echo
-echo "Avancement: 0% ---------------------------------------------------------------------------------------------------> Environnement "
-echo
+echo $(date)
+echo "***"
+echo "*** Updating packages list"
+echo "***"
+echo 5 > ${PROGRESS_FILE}
+apt-get clean
+echo 10 > ${PROGRESS_FILE}
+apt-get update
+echo 20 > ${PROGRESS_FILE}
 
-echo 50 > ${PROGRESS_FILE}
-echo
-echo "Avancement: 50% ---------------------------------------------------------------------------------------------------> Ajout au groupe dialout."
-echo
+echo "***"
+echo "*** Installing required packages"
+echo "***"
+apt-get install -y python3
+echo 60 > ${PROGRESS_FILE}
 
+# Tcharp38: is this part still required ?
 echo "**Ajout du user www-data dans le groupe dialout (accès à la zigate)**"
 if [[ `groups www-data | grep -c dialout` -ne 1 ]]; then
     useradd -g dialout www-data
@@ -68,8 +42,8 @@ if [[ `groups www-data | grep -c dialout` -ne 1 ]]; then
  fi
 
 echo 100 > ${PROGRESS_FILE}
-echo
-echo "Avancement: 100% ---------------------------------------------------------------------------------------------------> FIN"
-echo
-
+echo $(date)
+echo "***"
+echo "*** Installing ended"
+echo "***"
 rm ${PROGRESS_FILE}