+++ docs/fr_FR/Changelog.rst
@@ -3,6 +3,7 @@ ChangeLog
 
 - Page maintenance: Infos clefs: Mise-à-jour.
 - Interne: Mise-à-jour 'info.json'. Version minimale du core = 4.0.
+- Dependances: Mise-à-jour du code. Unique dépendance de base = python3.
 
 230718-BETA-1
 -------------